Move dict functions to own file and remove final cyclical imports
